feat: move `Header` to module
